Change filtering to use a common dictionary and rely less on frequency.
Some common words still have relatively low frequency in the Wikipedia corpus. Instead of restricting to a >=2,000 frequency words on Wikipedia, we will start with a more common dictionary of words (/usr/share/dict/words, in this case), join it with the frequency list, and eliminate words with _very_ low frequency (<300 by default for now).

This gives a smaller initial word list, but seems to contain a better candidate set, since it includes common words that also however, less frequent on Wikipedia.