Add a .go program to filter a frequency file based on dictionary words.
The point here is to take a frequency file that is very liberal in what it defines as a word, and reduce it to things that are in a more conservative dictionary file. But since I'm not sure whether wordle uses plurals and such (which may not be in the dictionary file), we add in a stemming step to try to keep various suffixes.